jarvis_assistant/requirements.txt
# Core AI
openai>=1.0.0
python-dotenv>=1.0.0

# Voice
SpeechRecognition>=3.10.0
edge-tts>=6.1.9
pyaudio>=0.2.14
pygame>=2.5.2
pvporcupine>=3.0.0
openwakeword>=0.6.0

# Utilities
requests>=2.31.0
aiohttp>=3.9.0
rich>=13.0.0  # Beautiful terminal output
psutil>=5.9.0  # System monitoring
python-miio>=0.5.12  # Xiaomi Smart Home
micloud>=0.6  # Xiaomi Cloud Token Extractor
yeelight>=0.7.13 # LAN Control for Yeelights (No Token needed)

# Optional: Smart Home
# homeassistant-api>=0.1.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0

# AEC & Audio
websockets

# Information Services
yfinance
feedparser
